+++ frontend/lost/src/containers/Workers/WorkersTable.jsx
@@ -1,115 +1,47 @@
-import React, { Component } from 'react'
-
+import { useEffect, useState } from 'react'
 import { connect } from 'react-redux'
-import { Badge, Button, Table } from 'reactstrap'
-// import Modal from 'react-modal'
-import LogModal from '../../components/LogModal'
+import { Badge, Table } from 'reactstrap'
 import actions from '../../actions'
 
 const { getWorkers, getWorkerLogFile } = actions
 
-const customStyles = {
-    content: {
-        top: '50%',
-        left: '50%',
-        right: 'auto',
-        bottom: 'auto',
-        marginRight: '-50%',
-        transform: 'translate(-50%, -50%)',
-        width: '85%',
-        height: '85%',
-        maxWidth: '75rem',
-    },
-    overlay: {
-        backgroundColor: 'rgba(0,0,0,0.75)',
-    },
-}
+const WorkersTable = ({ workers, getWorkers }) => {
+    const [modalsIsOpen, setModalsIsOpen] = useState([])
 
-class WorkersTable extends Component {
-    constructor(props) {
-        super(props)
-        this.state = {
-            logBlobUrl: '',
-            modalsIsOpen: [],
-        }
-    }
+    useEffect(() => {
+        getWorkers()
+        const workertimer = setInterval(() => getWorkers(), 5000)
 
-    componentDidMount() {
-        this.props.getWorkers()
-        this.workertimer = setInterval(() => this.props.getWorkers(), 5000)
-    }
+        return () => clearInterval(workertimer)
+    }, [getWorkers])
 
-    handleLogfileButtonClick(worker) {
-        this.setState({
-            modalsIsOpen: this.state.modalsIsOpen.map((el) => ({
-                idx: el.idx,
-                isOpen: el.idx === worker.idx ? true : false,
-            })),
-        })
-    }
-
-    componentWillUnmount() {
-        clearInterval(this.workertimer)
-        this.workertimer = null
-    }
-
-    componentWillReceiveProps() {
-        if (this.state.modalsIsOpen.length !== this.props.workers.length) {
-            this.setState({
-                modalsIsOpen: this.props.workers.map((el) => ({
-                    idx: el.idx,
+    useEffect(() => {
+        if (modalsIsOpen.length !== workers.length) {
+            setModalsIsOpen(
+                workers.map((worker) => ({
+                    idx: worker.idx,
                     isOpen: false,
                 })),
-            })
-        }
-    }
-
-    renderLogFileModal(worker) {
-        if (this.state.modalsIsOpen.length) {
-            const isOpen = this.state.modalsIsOpen.filter(
-                (el) => el.idx === worker.idx,
-            )[0].isOpen
-            return (
-                <LogModal
-                    actionType={LogModal.TYPES.WORKERS}
-                    isDownloadable
-                    wiLogId={worker.idx}
-                    isOpen={isOpen}
-                    toggle={() => {
-                        this.setState({
-                            modalsIsOpen: this.state.modalsIsOpen.map((el) => {
-                                if (el.idx === worker.idx) {
-                                    return {
-                                        ...el,
-                                        isOpen: !el.isOpen,
-                                    }
-                                }
-                                return el
-                            }),
-                        })
-                    }}
-                    logPath={isOpen ? `${worker.worker_name}.log` : null}
-                />
             )
         }
-        return null
-    }
+    }, [workers, modalsIsOpen.length])
 
-    renderTableBody() {
+    const renderTableBody = () => {
         return (
             <tbody>
-                {this.props.workers.map((worker) => {
+                {workers.map((worker) => {
                     let statusColor = 'success'
                     let statusText = 'Online'
 
                     if (worker.timestamp) {
-                        const someSecondsAgo = new Date(Date.now() - 15000) // fix this to correct timestamp
+                        const someSecondsAgo = new Date(Date.now() - 15000)
                         const lastActivityDate = new Date(worker.timestamp)
                         if (lastActivityDate < someSecondsAgo) {
                             statusColor = 'danger'
                             statusText = 'Offline'
                         }
                     }
+
                     return (
                         <tr key={worker.idx}>
                             <td className="text-center">
@@ -138,40 +70,29 @@ class WorkersTable extends Component {
                             <td className="text-center">
                                 <div>{worker.in_progress}</div>
                             </td>
-                            {/* <td className="text-center">
-                                <Button
-                                    onClick={() => this.handleLogfileButtonClick(worker)}
-                                >
-                                    Logs
-                                </Button>
-                                {this.renderLogFileModal(worker)}
-                            </td> */}
                         </tr>
                     )
                 })}
             </tbody>
         )
     }
-    render() {
-        return (
-            <Table hover responsive className="table-outline mb-0 d-none d-sm-table">
-                <thead className="thead-light">
-                    <tr>
-                        <th className="text-center">Name</th>
-                        <th className="text-center">Environment</th>
-                        <th className="text-center">Status</th>
-                        <th className="text-center">Resources</th>
-                        <th className="text-center">Jobs</th>
-                        {/* <th className="text-center">Logs</th> */}
-                    </tr>
-                </thead>
-                {this.renderTableBody()}
-            </Table>
-        )
-    }
+
+    return (
+        <Table hover responsive className="table-outline mb-0 d-none d-sm-table">
+            <thead className="thead-light">
+                <tr>
+                    <th className="text-center">Name</th>
+                    <th className="text-center">Environment</th>
+                    <th className="text-center">Status</th>
+                    <th className="text-center">Resources</th>
+                    <th className="text-center">Jobs</th>
+                </tr>
+            </thead>
+            {renderTableBody()}
+        </Table>
+    )
 }
-function mapStateToProps(state) {
-    return { workers: state.worker.workers }
-}
+
+const mapStateToProps = (state) => ({ workers: state.worker.workers })
 
 export default connect(mapStateToProps, { getWorkers, getWorkerLogFile })(WorkersTable)