+++ frontend/lost/src/containers/pipeline/start/2/modals/types/LoopModal.jsx
@@ -1,48 +1,44 @@
-import React, { Component } from 'react'
+import { useCallback } from 'react'
+import { connect } from 'react-redux'
+import { Input } from 'reactstrap'
+import actions from '../../../../../../actions/pipeline/pipelineStartModals/loop'
 import Table from '../../../../globalComponents/modals/Table'
 
-// import actions from "actions/pipeline/pipelineStartModals/loop";
-import actions from '../../../../../../actions/pipeline/pipelineStartModals/loop'
-import { Input } from 'reactstrap'
-import { connect } from 'react-redux'
 const { inputMaxIteration } = actions
 
-class LoopModal extends Component {
-    constructor() {
-        super()
-        this.onInput = this.onInput.bind(this)
-    }
-    onInput(e) {
-        let number = Number(e.target.value)
-        if (typeof number === 'number') {
-            this.props.inputMaxIteration(this.props.peN, number)
-        }
-    }
-    render() {
-        return (
-            <Table
-                data={[
-                    {
-                        key: 'Max Iteration',
-                        value:
-                            typeof this.props.exportData.loop.maxIteration ===
-                            'number' ? (
-                                <Input
-                                    min={-1}
-                                    onInput={this.onInput}
-                                    defaultValue={this.props.exportData.loop.maxIteration}
-                                    placeholder="Amount"
-                                    type="number"
-                                    step="1"
-                                />
-                            ) : (
-                                'Infinity'
-                            ),
-                    },
-                ]}
-            />
-        )
-    }
+const LoopModal = ({ exportData, inputMaxIteration, peN }) => {
+    const onInput = useCallback(
+        (e) => {
+            const number = Number(e.target.value)
+            if (!isNaN(number)) {
+                inputMaxIteration(peN, number)
+            }
+        },
+        [inputMaxIteration, peN],
+    )
+
+    return (
+        <Table
+            data={[
+                {
+                    key: 'Max Iteration',
+                    value:
+                        typeof exportData.loop.maxIteration === 'number' ? (
+                            <Input
+                                min={-1}
+                                onInput={onInput}
+                                defaultValue={exportData.loop.maxIteration}
+                                placeholder="Amount"
+                                type="number"
+                                step="1"
+                            />
+                        ) : (
+                            'Infinity'
+                        ),
+                },
+            ]}
+        />
+    )
 }
 
 export default connect(null, { inputMaxIteration })(LoopModal)