+++ frontend/lost/src/containers/pipeline/start/2/modals/types/ScriptModal.jsx
@@ -1,58 +1,54 @@
-import React, { Component } from 'react'
-import Table from '../../../../globalComponents/modals/Table'
-
-import ArgumentsTable from '../../../../globalComponents/modals/ScriptArgumentsTable'
-
-import CollapseCard from '../../../../globalComponents/modals/CollapseCard'
-
+import { useCallback } from 'react'
 import { connect } from 'react-redux'
 import actions from '../../../../../../actions/pipeline/pipelineStartModals/script'
+import CollapseCard from '../../../../globalComponents/modals/CollapseCard'
+import ArgumentsTable from '../../../../globalComponents/modals/ScriptArgumentsTable'
+import Table from '../../../../globalComponents/modals/Table'
 
 const { updateArguments } = actions
-class ScriptModal extends Component {
-    constructor() {
-        super()
-        this.argumentTableOnInput = this.argumentTableOnInput.bind(this)
-    }
-    argumentTableOnInput(e) {
-        let arg = this.props.exportData.script.arguments
-        const key = e.target.getAttribute('data-ref')
-        const value = e.target.value
-        arg[key].value = value
-        this.props.updateArguments(this.props.peN, arg)
-    }
-    render() {
-        return (
-            <>
+
+const ScriptModal = ({ exportData, peN, updateArguments }) => {
+    const argumentTableOnInput = useCallback(
+        (e) => {
+            const arg = { ...exportData.script.arguments }
+            const key = e.target.getAttribute('data-ref')
+            const value = e.target.value
+            arg[key].value = value
+            updateArguments(peN, arg)
+        },
+        [exportData.script.arguments, peN, updateArguments],
+    )
+
+    return (
+        <>
+            <Table
+                data={[
+                    {
+                        key: 'Script Name',
+                        value: exportData.script.name,
+                    },
+                    {
+                        key: 'Description',
+                        value: exportData.script.description,
+                    },
+                ]}
+            />
+            <ArgumentsTable
+                onInput={argumentTableOnInput}
+                data={exportData.script.arguments}
+            />
+            <CollapseCard>
                 <Table
                     data={[
                         {
-                            key: 'Script Name',
-                            value: this.props.exportData.script.name,
-                        },
-                        {
-                            key: 'Description',
-                            value: this.props.exportData.script.description,
+                            key: 'Path',
+                            value: exportData.script.path,
                         },
                     ]}
                 />
-                <ArgumentsTable
-                    onInput={this.argumentTableOnInput}
-                    data={this.props.exportData.script.arguments}
-                />
-                <CollapseCard>
-                    <Table
-                        data={[
-                            {
-                                key: 'Path',
-                                value: this.props.exportData.script.path,
-                            },
-                        ]}
-                    />
-                </CollapseCard>
-            </>
-        )
-    }
+            </CollapseCard>
+        </>
+    )
 }
 
 export default connect(null, { updateArguments })(ScriptModal)